requirements.txt
anthropic==0.84.0
python-docx==1.2.0
pdfplumber==0.11.9
flask==3.1.3
python-dotenv==1.2.2
gunicorn==21.2.0
reportlab
psycopg2-binary==2.9.9